# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Providers & Models Catalog
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PricingType = typeof PricingType[keyof typeof PricingType];


export const PricingType = {
  free: 'free',
  paid: 'paid',
  freemium: 'freemium',
  open_source: 'open_source',
} as const;

export type Modality = typeof Modality[keyof typeof Modality];


export const Modality = {
  text: 'text',
  image: 'image',
  audio: 'audio',
  video: 'video',
  code: 'code',
  agents: 'agents',
} as const;

export interface Provider {
  id: number;
  slug: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  pricingUrl?: string | null;
  /** @nullable */
  docsUrl?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  isOpenSource: boolean;
  /** @nullable */
  headquarters?: string | null;
  modelCount: number;
  lastVerifiedAt: string;
  verifiedByAdmin?: boolean;
  createdAt: string;
}

export interface Model {
  id: number;
  providerId: number;
  slug: string;
  name: string;
  /** @nullable */
  description?: string | null;
  pricingType: PricingType;
  /** @nullable */
  priceInputPerMillionTokens?: number | null;
  /** @nullable */
  priceOutputPerMillionTokens?: number | null;
  /** @nullable */
  priceUnit?: string | null;
  /** @nullable */
  priceNotes?: string | null;
  /** @nullable */
  contextWindowTokens?: number | null;
  /** @nullable */
  maxOutputTokens?: number | null;
  modalitiesInput: Modality[];
  modalitiesOutput: Modality[];
  specialties: string[];
  /** @nullable */
  usageLimits?: string | null;
  apiAvailable: boolean;
  /** @nullable */
  officialLink?: string | null;
  /** @nullable */
  releaseDate?: string | null;
  lastUpdatedAt: string;
  verifiedByAdmin?: boolean;
  isPotentiallyOutdated: boolean;
  createdAt: string;
}

export type ProviderDetail = Provider & {
  models: Model[];
};

export interface ProviderInput {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  name: string;
  description?: string;
  websiteUrl?: string;
  pricingUrl?: string;
  docsUrl?: string;
  logoUrl?: string;
  isOpenSource: boolean;
  headquarters?: string;
  lastVerifiedAt?: string;
}

export interface ProviderUpdate {
  /** @minLength 1 */
  slug?: string;
  /** @minLength 1 */
  name?: string;
  description?: string;
  websiteUrl?: string;
  pricingUrl?: string;
  docsUrl?: string;
  logoUrl?: string;
  isOpenSource?: boolean;
  headquarters?: string;
  lastVerifiedAt?: string;
}

export type ModelWithProvider = Model & ({
  providerName: string;
  providerSlug: string;
  providerIsOpenSource: boolean;
  /** @nullable */
  providerLogoUrl?: string | null;
});

export interface ModelInput {
  providerId: number;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  name: string;
  description?: string;
  pricingType: PricingType;
  priceInputPerMillionTokens?: number;
  priceOutputPerMillionTokens?: number;
  priceUnit?: string;
  priceNotes?: string;
  contextWindowTokens?: number;
  maxOutputTokens?: number;
  modalitiesInput: Modality[];
  modalitiesOutput: Modality[];
  specialties?: string[];
  usageLimits?: string;
  apiAvailable: boolean;
  officialLink?: string;
  releaseDate?: string;
  lastUpdatedAt?: string;
}

export interface ModelUpdate {
  providerId?: number;
  /** @minLength 1 */
  slug?: string;
  /** @minLength 1 */
  name?: string;
  description?: string;
  pricingType?: PricingType;
  priceInputPerMillionTokens?: number;
  priceOutputPerMillionTokens?: number;
  priceUnit?: string;
  priceNotes?: string;
  contextWindowTokens?: number;
  maxOutputTokens?: number;
  modalitiesInput?: Modality[];
  modalitiesOutput?: Modality[];
  specialties?: string[];
  usageLimits?: string;
  apiAvailable?: boolean;
  officialLink?: string;
  releaseDate?: string;
  lastUpdatedAt?: string;
}

export interface PricingBreakdown {
  pricingType: PricingType;
  count: number;
}

export interface ModalityBreakdown {
  modality: Modality;
  count: number;
}

export interface StatsSummary {
  totalProviders: number;
  totalModels: number;
  openSourceProviders: number;
  pricingBreakdown: PricingBreakdown[];
  modalityBreakdown: ModalityBreakdown[];
  recentlyUpdatedModels: ModelWithProvider[];
  potentiallyOutdatedCount: number;
}

export type ListProvidersParams = {
search?: string;
isOpenSource?: boolean;
sort?: ListProvidersSort;
};

export type ListProvidersSort = typeof ListProvidersSort[keyof typeof ListProvidersSort];


export const ListProvidersSort = {
  name_asc: 'name_asc',
  name_desc: 'name_desc',
  recently_updated: 'recently_updated',
  model_count_desc: 'model_count_desc',
} as const;

export type ListModelsParams = {
search?: string;
providerId?: number;
/**
 * Comma-separated list, e.g. free,paid,freemium,open_source
 */
pricingType?: string;
/**
 * Comma-separated list of modalities
 */
modality?: string;
specialty?: string;
isOpenSource?: boolean;
apiAvailable?: boolean;
sort?: ListModelsSort;
};

export type ListModelsSort = typeof ListModelsSort[keyof typeof ListModelsSort];


export const ListModelsSort = {
  name_asc: 'name_asc',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  context_desc: 'context_desc',
  recently_updated: 'recently_updated',
} as const;