# lib/api-zod/src/generated/types/model.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Providers & Models Catalog
 * OpenAPI spec version: 0.1.0
 */
import type { Modality } from './modality';
import type { PricingType } from './pricingType';

export interface Model {
  id: number;
  providerId: number;
  slug: string;
  name: string;
  /** @nullable */
  description?: string | null;
  pricingType: PricingType;
  /** @nullable */
  priceInputPerMillionTokens?: number | null;
  /** @nullable */
  priceOutputPerMillionTokens?: number | null;
  /** @nullable */
  priceUnit?: string | null;
  /** @nullable */
  priceNotes?: string | null;
  /** @nullable */
  contextWindowTokens?: number | null;
  /** @nullable */
  maxOutputTokens?: number | null;
  modalitiesInput: Modality[];
  modalitiesOutput: Modality[];
  specialties: string[];
  /** @nullable */
  usageLimits?: string | null;
  apiAvailable: boolean;
  /** @nullable */
  officialLink?: string | null;
  /** @nullable */
  releaseDate?: string | null;
  lastUpdatedAt: Date;
  verifiedByAdmin?: boolean;
  isPotentiallyOutdated: boolean;
  createdAt: Date;
}