# lib/api-zod/src/generated/types/modelUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Providers & Models Catalog
 * OpenAPI spec version: 0.1.0
 */
import type { Modality } from './modality';
import type { PricingType } from './pricingType';

export interface ModelUpdate {
  providerId?: number;
  /** @minLength 1 */
  slug?: string;
  /** @minLength 1 */
  name?: string;
  description?: string;
  pricingType?: PricingType;
  priceInputPerMillionTokens?: number;
  priceOutputPerMillionTokens?: number;
  priceUnit?: string;
  priceNotes?: string;
  contextWindowTokens?: number;
  maxOutputTokens?: number;
  modalitiesInput?: Modality[];
  modalitiesOutput?: Modality[];
  specialties?: string[];
  usageLimits?: string;
  apiAvailable?: boolean;
  officialLink?: string;
  releaseDate?: string;
  lastUpdatedAt?: Date;
}