# lib/api-zod/src/generated/types/pricingBreakdown.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Providers & Models Catalog
 * OpenAPI spec version: 0.1.0
 */
import type { PricingType } from './pricingType';

export interface PricingBreakdown {
  pricingType: PricingType;
  count: number;
}