# lib/api-zod/src/generated/types/providerDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Providers & Models Catalog
 * OpenAPI spec version: 0.1.0
 */
import type { Model } from './model';
import type { Provider } from './provider';

export type ProviderDetail = Provider & {
  models: Model[];
};